backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
onnxruntime==1.19.2
numpy==1.26.4
Pillow==10.4.0
requests==2.32.5
jinja2==3.1.4
aiofiles==24.1.0
python-multipart==0.0.9
pytesseract==0.3.13
httpx==0.27.0
